requirements.txt
strands-agents
bedrock-agentcore
pandas
beautifulsoup4